linux: Don't always call lx_close after mmapping a dataspace
lx_close on the dataspace file handle must only be called if outside
the core.  Within core lx_close on that file can cause the file handle
to be reused for other purposes (such as an answer socket for IPC).

The file handle is still stored within the dataspace and the lx_close
call from within Ram_session_component::_revoke_ram_ds() will happily
close the answer socket.

Implementing the suggestion of Norman Feske to provide a method named
_release_dataspace which is a no-op within core and lx_close the file
out-side of core.